Added selections, improved event handling, + more
Selections - added back in selection capabilities. Have not seen any
issues with it

Event handling - several improvements have been made:
1) Both the text label and the arcs can be hovered or clicked for
events. Previously, the text label on the arcs would not handle events
2) The click event now removes the mouseover events, so that the
brushing stays active as the user decides whether to confirm the
selection or not
3) The click event brushing is additive, so as you click additional
selections inside the chart, the new selections come into the foreground
with the previous selections

Other improvements: the tooltip div was being created every time the
extension rendered, resulting in many tooltip divs generating in the
background. Now the tooltip div is only created if it doesn't already
exist in the document; otherwise the existing tooltip div is used.